requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
bitarray==2.9.2
cached-property==1.5.2
certifi==2024.2.2
charset-normalizer==3.3.2
colorlog==6.8.2
cytoolz==0.12.3
eth-account==0.11.0
eth-bloom==3.0.0
eth-hash==0.7.0
eth-keyfile==0.8.0
eth-keys==0.4.0
eth-rlp==1.0.1
eth-tester==0.9.1b2
eth-typing==3.5.2
eth-utils==2.3.1
eth_abi==5.0.0
frozenlist==1.4.1
hexbytes==0.3.1
idna==3.6
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lru-dict==1.2.0
multidict==6.0.5
mypy-extensions==1.0.0
parsimonious==0.9.0
protobuf==4.25.3
py-ecc==6.0.0
py-evm==0.7.0a4
py-geth==4.3.0
pycryptodome==3.20.0
pyethash==0.1.27
python-dotenv==1.0.1
pyunormalize==15.1.0
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rlp==3.0.0
rpds-py==0.18.0
safe-pysha3==1.0.4
semantic-version==2.10.0
setuptools==69.1.1
sortedcontainers==2.4.0
toolz==0.12.1
trie==2.2.0
typing_extensions==4.10.0
urllib3==2.2.1
web3==6.15.1
websockets==12.0
yarl==1.9.4
